option_trading/requirements.txt
box2d=2.3.10=pypi_0
conda=4.10.1=py37haa95532_1
cudatoolkit=10.2.89=h74a9793_1
gym=0.18.3=pypi_0
ipython=7.13.0=py37h5ca1d4c_0
matplotlib=3.1.3=py37_0
numpy=1.19.2=py37hadc3359_0
pandas=1.0.3=py37h47e9c7a_0
python=3.7.7=h81c818b_4
pytorch=1.7.1=py3.7_cuda102_cudnn7_0
scikit-learn=0.22.1=py37h6288b17_0
scipy=1.6.2=py37h14eb087_0
seaborn=0.11.1=pyhd3eb1b0_0
torchaudio=0.7.2=py37
torchvision=0.8.2=py37_cu102
